Add separate function generateFilename, rename getter, add comments to constructor
