+++ lib/filename-generators/index.js
