+++ lib/scraper.js
@@ -12,7 +12,7 @@ var defaults = require('./config/defaults');
 var recursiveSources = require('./config/recursive-sources');
 var Resource = require('./resource');
 
-var getFilenameGenerator = require('./filename-generators/filename-generator-getter');
+var getFilenameGenerator = require('./filename-generators');
 var makeRequest = require('./request');
 var normalizeUrl = require('normalize-url');
 
@@ -27,33 +27,35 @@ function loadHtmlAndCss (context, po) {
 function Scraper (options) {
 	var self = this;
 
+	// Extend options
 	self.options = _.extend({}, defaults, options);
 	self.options.request = _.extend({}, defaults.request, options.request);
-	self.options.filenameGenerator = getFilenameGenerator(self.options.filenameGenerator);
-
-	// Create makeRequest function with custom request params
-	self.makeRequest = makeRequest.bind(null, self.options.request);
+	self.options.urls = _.isArray(self.options.urls) ? self.options.urls : [self.options.urls];
 
 	if (self.options.directory) {
 		self.options.absoluteDirectoryPath = path.resolve(process.cwd(), self.options.directory);
 	}
 
-	// Create array of Resource for downloading
-	self.options.urls = _.isArray(self.options.urls) ? self.options.urls : [self.options.urls];
+	if (self.options.recursive) {
+		self.options.sources = _.union(self.options.sources, recursiveSources);
+	}
+
+	// Create custom functions for request and filename generating
+	self.makeRequest = makeRequest.bind(null, self.options.request);
+	self.generateFilename = getFilenameGenerator(self.options.filenameGenerator);
+
+	// Custom structures
+	// Array of Resources for downloading
 	self.originalResources = _.map(self.options.urls, function createResource (obj) {
 		var url = _.isObject(obj) && _.has(obj, 'url') ? obj.url : obj;
 		var filename = _.isObject(obj) && _.has(obj, 'filename') ? obj.filename : self.options.defaultFilename;
 		return new Resource(url, filename);
 	});
 
-	if (self.options.recursive) {
-		self.options.sources = _.union(self.options.sources, recursiveSources);
-	}
-
-	self.occupiedFileNames = [];
-	self.loadedResources = [];
-	self.respondedResourcePromises = {};
-	self.loadedResourcePromises = {};
+	self.occupiedFileNames = []; // Array of unavailable filenames
+	self.loadedResources = [];   // Array of loaded Resources
+	self.respondedResourcePromises = {}; // Map url -> request promise
+	self.loadedResourcePromises = {};    // Map url -> save-to-fs promise
 }
 
 Scraper.prototype.addOccupiedFileName = function addOccupiedFileName (filename) {
@@ -161,7 +163,7 @@ Scraper.prototype.requestResource = function requestResource (resource) {
 		// Url may be changed in redirects
 		resource.setUrl(data.url);
 		
-		var filename = self.options.filenameGenerator(resource, self.options, self.getOccupiedFileNames());
+		var filename = self.generateFilename(resource, self.options, self.getOccupiedFileNames());
 		resource.setFilename(filename);
 		self.addOccupiedFileName(filename);
 		self.addRespondedResourcePromise(data.url, promise);
